% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{latlon}
\alias{latlon}
\title{Making a linear regression function using S3 class}
\usage{
latlon(address, apiKey)
}
\arguments{
\item{address}{as vector}

\item{apikey}{as character}
}
\value{
vector
}
\description{
Making a linear regression function using S3 class
}
\examples{
latlon(address = "New Delhi, India", apiKey = "randomenumbersandletters")
}
