% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{reverselatlon}
\alias{reverselatlon}
\title{Function to make api call using Google apis and adress for provided latitude, longitude values}
\usage{
reverselatlon(lat, lon, apiKey)
}
\arguments{
\item{lat}{as number}

\item{lon}{as number}

\item{apiKey}{as character}
}
\value{
character
}
\description{
Function to make api call using Google apis and adress for provided latitude, longitude values
}
