% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{latlon}
\alias{latlon}
\title{Function to make api call using Google apis and return latitude, longitude for provided address}
\usage{
latlon(address, apiKey)
}
\arguments{
\item{address}{as vector}

\item{apiKey}{as character}
}
\value{
vector
}
\description{
Function to make api call using Google apis and return latitude, longitude for provided address
}
\examples{
latlon(address = "New Delhi, India", apiKey = "randomenumbersandletters")
}
